Updated everything - see description
The math workings are in the appendix for now to save word count.

I am busy working on explaining the model and discussing the results.

I am struggling to include symbols in the tables I made to explain the game and in the simultaneous game. This will have to be changed.

Any images should be put in the img folder (except the front page one which is in the tex folder).